Update PHP version & fix tests
